+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/jwt/controller/AuthenticationController.java
@@ -3,7 +3,6 @@ package codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.controller.response.ApiResponse;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.request.LoginRequest;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.request.RefreshTokenRequest;
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.request.SignUpRequest;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.response.JwtResponse;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.service.JwtService;
 import javax.servlet.http.HttpServletRequest;
@@ -16,19 +15,13 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
 @RestController
-public class JwtController {
+public class AuthenticationController {
 
     private final JwtService jwtService;
 
     @PostMapping("/api/login")
     public JwtResponse login(@RequestBody @Valid LoginRequest request) {
         return JwtResponse.from(jwtService.login(LoginRequest.to(request)));
-    }
-
-    @PostMapping("/api/signup")
-    public ApiResponse signUp(@RequestBody @Valid SignUpRequest request) {
-        jwtService.signUp(SignUpRequest.to(request));
-        return ApiResponse.success(HttpStatus.OK);
     }
 
     @PostMapping("/api/auth/reissue")
