+++ be/src/test/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/jwt/integration/JwtIntegrationTest.java
@@ -7,7 +7,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import codesquad.kr.gyeonggidoidle.issuetracker.annotation.IntegrationTest;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.request.LoginRequest;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.request.RefreshTokenRequest;
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.controller.request.SignUpRequest;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.entity.Jwt;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.jwt.entity.JwtProvider;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -15,7 +14,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Map;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Order;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
@@ -35,7 +33,6 @@ public class JwtIntegrationTest {
 
     @DisplayName("이메일과 비밀번호를 받아 JwtResponse를 반환한다.")
     @Test
-    @Order(1)
     void login() throws Exception {
         // given
         LoginRequest request = LoginRequest.builder()
@@ -58,30 +55,8 @@ public class JwtIntegrationTest {
                 );
     }
 
-    @DisplayName("이메일, 비밀번호, 프로필사진을 받아 회원가입을 한다.")
-    @Test
-    @Order(2)
-    void signUp() throws Exception {
-        // given
-        SignUpRequest request = SignUpRequest.builder()
-                .email("test@test.com")
-                .password("!test1234")
-                .profile("profile1234")
-                .build();
-
-        // when
-        ResultActions resultActions = mockMvc.perform(post("/api/signup")
-                .contentType(MediaType.APPLICATION_JSON)
-                .content(toJson(request)));
-
-        // then
-        resultActions
-                .andExpect(status().isOk());
-    }
-
     @DisplayName("refreshToken을 받아서 새로운 accessToken이 담긴 JwtResponse를 반환한다.")
     @Test
-    @Order(3)
     @Disabled
     void reissueAccessToken() throws Exception {
         // given
@@ -105,7 +80,6 @@ public class JwtIntegrationTest {
 
     @DisplayName("HttpServletRequest에 담긴 claim의 memberId를 가지고 refreshToken을 삭제한다.")
     @Test
-    @Order(4)
     void logout() throws Exception {
         // given
         Jwt jwt = makeToken();
